+++ src/i18n.js
@@ -9,7 +9,6 @@ function addTranslations(translationsObj) {
 function translate(key, params = {}) {
   const paths = key.split('.');
   let translation = translations[paths[0]];
-  console.log(translation);
   for (let i = 1; i < paths.length; i++) {
     translation = translation[paths[i]];
   }